Auto stash before merge of "VHung" and "origin/PHung"

+++ SpringBoot-Thymeleaf-JPA-ALOHCMUTE/src/main/java/hcmute/alohcmute/repositories/BaiVietRepository.java
@@ -6,6 +6,11 @@ import org.springframework.stereotype.Repository;
 import hcmute.alohcmute.entities.BaiViet;
 
 @Repository
-public interface BaiVietRepository extends JpaRepository<BaiViet, Long> {
+public interface BaiVietRepository extends JpaRepository<BaiViet, Integer> {
 
+	
+
+	
+	
 }
+
